Move to root for types/eslint
