+++ .storybook/preview.js
@@ -1,4 +1,4 @@
-import { isTestRunner } from '../dist/esm/is-test-runner/is-test-runner';
+import { isTestRunner } from '../is-test-runner';
 
 const withSkippableTests = (StoryFn, { parameters }) => {
   if (parameters.test?.skip && isTestRunner()) {
