+++ is-test-runner.d.ts
@@ -1,0 +1,1 @@
+export declare const isTestRunner: () => boolean;
